Added model folder and moved model files into it
